Added test case for Admin remove item for user funcitonality

+++ src/main/java/com/wishlist/cst438project2/controller/AdminController.java
@@ -189,7 +189,7 @@ public class AdminController {
      * returns timestamp of successful deletion
      */
     @DeleteMapping("/removeItems")
-    public String removeItem(@RequestHeader String accessToken, @RequestParam String itemName, @RequestParam int userId) {
+    public String removeItem(@RequestHeader String accessToken, @RequestParam String itemName, @RequestParam Long userId) {
 
         UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
