+++ src/main/java/com/wishlist/cst438project2/service/AdminService.java
@@ -22,7 +22,7 @@ public interface AdminService {
 
     UserLoginDTO login(SignInDTO signInDTO);
 
-    String removeItem(String itemName, int userId);
+    String removeItem(String itemName, Long userId);
 
     UserDTO updateUser(UserDTO userDTO);
 }
